Many compiler fixes    Changed equal to equalString where appropriate    Made many conversions explicit    Removed semis at end of method headers    Changed some calculated properties to readonly    Added hash functions to objects defining isEqual (Download, TivoShow); not expected to be used.    Removed possible nil returns in Tasks to avoid compiler warning on including in array    Moved nil checks from x?x:y to x?:y    Avoid modifying parameters    In Subscription, fixed 2 bugs referencing NSNumber as bools    In ProgramTableView, fixed wrong enum reference (irrelevant as they map to same number)    In FormatEditorController, isolated ignoring SelectorLeak warning to method that needs it    Ensure that delegates are nilled in dealloc    Removed unused MTGradientView    Added compiler checks: Wall", 					"-Wextra",  (but ignored "-Wunused-parameter" in cTiVo-Prefix.pch 					"-Wnewline-eof", 					"-Wshadow", 					"-Wundef", 					"-Wdeprecated-implementations",
Changed MaxProgressDelay to preference item (seconds)
Added reloading of tvdbCache on each execution: note it never shrinks; add date last used?
Fixed not restarting after encode finishes.
Filter out Tivo Desktop
Added "language=all" to tvdb checks to pickup Spanish (and presumably other) titles
Moved code-sig identity to allow others to build

+++ cTiVo/MTPreferencesWindowController.m
@@ -108,8 +108,8 @@
 	NSSize viewSize = ((NSViewController *)tabViewItem.windowController).view.frame.size;
 	NSSize newSize = NSMakeSize(viewSize.width+40, viewSize.height + 100);
 	NSRect frame = self.window.frame;
-	float newXOrigin = frame.origin.x + (frame.size.width - newSize.width)/2.0;
-	float newYOrigin = frame.origin.y + frame.size.height - newSize.height;
+	double newXOrigin = frame.origin.x + (frame.size.width - newSize.width)/2.0;
+	double newYOrigin = frame.origin.y + frame.size.height - newSize.height;
 	return NSMakeRect(newXOrigin, newYOrigin, newSize.width, newSize.height);
 }
